// src/enums/classes.ts
/**
 * For global style classes - use component css module for component styles
 */
export enum Classes {
  LinkButton = 'link-button',
  ListStyle = 'list-style',
  InputContainer = 'input-container',
  CenterContent = 'center-content',
  BaseGutters = 'base-gutters',
  BasePadding = 'base-padding',
  CropImageContainer = 'crop-image-container',
  InputErrorState = 'input-error-state'
}